requirements.txt
flask
flask_login
markdown
passlib
pycryptodome
gunicorn
bleach